trial service module added
